Refactor

+++ CoinbasePro.Data/AutoMapper/AccountBalanceMapperProfile.cs
@@ -1,6 +1,7 @@
+using System;
 using AutoMapper;
-using CoinbasePro.Core.Dto.Data;
-using CoinbasePro.Core.Entities;
+using CoinbasePro.Data.Entities;
+using Hub.Shared.DataContracts.Banking;
 
 namespace CoinbasePro.Data.AutoMapper
 {
@@ -10,7 +11,7 @@ namespace CoinbasePro.Data.AutoMapper
         {
             CreateMap<AccountBalance, AccountBalanceDto>()
                 .ForMember(dest => dest.Balance, 
-                    opt => opt.MapFrom(x => x.Value))
+                    opt => opt.MapFrom(x => (decimal)x.Value))
                 .ReverseMap();
         }
     }
